Add Coordinator and Mono Agent
The Coordinator is the server that stores the differents revisions binaries, the benchmarks to run and the counters samples. The web interface is a work in progress and is not yet commited.

The Agent is the piece of software that is going to be run by a scheduler (Jenkins in our case) on regular interval/events (N times per day, on every M commits). It's job is simply to fetch the revision to run from the Coordinator, the configurations for the different benchmarks to run, run the benchmarks with the Log Profiler enabled, and send back the Log Profiler output back to the Coordinator.

The Monodroid and Monotouch agents are a work in progress and are going to use the Xamarin TestCloud platform to run on as many hardware devices as possible.